+++ tbomb/bomber.py
@@ -9,138 +9,7 @@ import argparse
 from datetime import datetime
 from concurrent.futures import ThreadPoolExecutor, as_completed
 
-
-try:
-    import requests
-    from colorama import Fore, Style
-except ImportError:
-    print("\tSome dependencies could not be imported (possibly not installed) ")
-    print("Type `pip3 install -r requirments.txt` to install all required packages")
-    sys.exit(1)
-
-class IconicDecorator(object):
-    def __init__(self):
-        self.PASS = Style.BRIGHT + Fore.GREEN + "[ ✔ ]" + Style.RESET_ALL
-        self.FAIL = Style.BRIGHT + Fore.RED + "[ ✘ ]" + Style.RESET_ALL
-        self.WARN = Style.BRIGHT + Fore.YELLOW + "[ ! ]" + Style.RESET_ALL
-        self.HEAD = Style.BRIGHT + Fore.CYAN + "[ # ]" + Style.RESET_ALL
-        self.CMDL = Style.BRIGHT + Fore.BLUE + "[ → ]" + Style.RESET_ALL
-        self.STDS = "     "
-
-class StatusDecorator(object):
-    def __init__(self):
-        self.PASS = Style.BRIGHT + Fore.GREEN + "[ SUCCESS ]" + Style.RESET_ALL
-        self.FAIL = Style.BRIGHT + Fore.RED + "[ FAILURE ]" + Style.RESET_ALL
-        self.WARN = Style.BRIGHT + Fore.YELLOW + "[ WARNING ]" + Style.RESET_ALL
-        self.HEAD = Style.BRIGHT + Fore.CYAN + "[ SECTION ]" + Style.RESET_ALL
-        self.CMDL = Style.BRIGHT + Fore.BLUE + "[ COMMAND ]" + Style.RESET_ALL
-        self.STDS = "           "
-
-class MessageDecorator(object):
-    def __init__(self, attr):
-        ICON = IconicDecorator()
-        STAT = StatusDecorator()
-        if attr == "icon":
-            self.PASS = ICON.PASS
-            self.FAIL = ICON.FAIL
-            self.WARN = ICON.WARN
-            self.HEAD = ICON.HEAD
-            self.CMDL = ICON.CMDL
-            self.STDS = ICON.STDS
-        elif attr == "stat":
-            self.PASS = STAT.PASS
-            self.FAIL = STAT.FAIL
-            self.WARN = STAT.WARN
-            self.HEAD = STAT.HEAD
-            self.CMDL = STAT.CMDL
-            self.STDS = STAT.STDS
-
-    def SuccessMessage(self, RequestMessage):
-        print(self.PASS + " " + Style.RESET_ALL + RequestMessage)
-
-    def FailureMessage(self, RequestMessage):
-        print(self.FAIL + " " + Style.RESET_ALL + RequestMessage)
-
-    def WarningMessage(self, RequestMessage):
-        print(self.WARN + " " + Style.RESET_ALL + RequestMessage)
-
-    def SectionMessage(self, RequestMessage):
-        print(self.HEAD + " " + Fore.CYAN + Style.BRIGHT + RequestMessage + Style.RESET_ALL)
-
-    def CommandMessage(self, RequestMessage):
-        return self.CMDL + " " + Style.RESET_ALL + RequestMessage
-
-    def GeneralMessage(self, RequestMessage):
-        print(self.STDS + " " + Style.RESET_ALL + RequestMessage)
-
-
-class APIProvider:
-
-    api_providers=[]
-
-    def __init__(self,cc,target,mode):    
-        with open('apidata.json', 'r') as file:
-            PROVIDERS = json.load(file)
-        self.config = None
-        self.cc = cc
-        self.target = target
-        self.mode = mode
-        self.index = 0
-        self.lock = threading.Lock()
-        providers=PROVIDERS.get(mode.lower(),{})
-        if (mode.lower()=="mail"):
-            APIProvider.api_providers = providers.get("mail",[])
-        else:
-            APIProvider.api_providers = providers.get(cc,providers.get("multi",[]))
-
-    def format(self):
-        config_dump = json.dumps(self.config)
-        config_dump = config_dump.replace("{target}",self.target).replace("{cc}",self.cc)
-        self.config = json.loads(config_dump)
-
-    def select_api(self):
-        try:
-            self.index = random.choice(range(len(APIProvider.api_providers)))
-        except IndexError:
-            self.index=-1
-        self.config = APIProvider.api_providers[self.index]
-        perma_headers = {"User-Agent":"Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:72.0) Gecko/20100101 Firefox/72.0"}
-        if "headers" in self.config:
-            self.config["headers"].update(perma_headers)
-        else:
-            self.config["headers"]=perma_headers
-        self.format()
-
-    def remove(self):
-        try:
-            del APIProvider.api_providers[self.index]
-            return True
-        except:
-            return False
-
-    def request(self):
-        self.select_api()
-        if not self.config or self.index==-1:
-            return None
-        identifier=self.config.pop("identifier","").lower()
-        del self.config['name']
-        self.config['timeout']=30
-        response=requests.request(**self.config)
-        return identifier in response.text.lower()
-
-    def hit(self):
-        try:
-            self.lock.acquire()
-            response = self.request()
-            if response==False:
-                self.remove()
-            return response
-        except:
-            response=False
-        finally:
-            self.lock.release()
-            return response
-
+from utils import *
 
 
 def readisdc():
@@ -150,7 +19,7 @@ def readisdc():
 
 def get_version():
     try:
-        return open(".version","r").read().strip()
+        return open(".version","r").read()
     except:
         return '1.0'
 
@@ -240,7 +109,7 @@ def update():
         do_zip_update()
 def check_for_updates():
     mesgdcrt.SectionMessage("Checking for updates")
-    fver = requests.get("https://raw.githubusercontent.com/TheSpeedX/TBomb/master/.version").text.strip()
+    fver = request.get("https://raw.githubusercontent.com/TheSpeedX/TBomb/master/.version").text.strip()
     if fver != __VERSION__:
         mesgdcrt.WarningMessage("An update is available")
         mesgdcrt.GeneralMessage("Starting update...")
@@ -391,11 +260,7 @@ if sys.version_info[0]!=3:
     mesgdcrt.FailureMessage("TBomb will work only in Python v3")
     sys.exit()
 
-try:
-    country_codes = readisdc()["isdcodes"]
-except FileNotFoundError:
-    update()
-
+country_codes = readisdc()["isdcodes"]
 mesgdcrt = MessageDecorator("icon")
 
 
@@ -430,7 +295,7 @@ if __name__ == "__main__":
     elif args.contributors:
         print("Contributors: "," ".join(__CONTRIBUTORS__))
     elif args.update:
-        update()
+        do_git_update()
     elif args.mail:
         selectnode(mode="mail")
     elif args.call:
